+++ .gitignore
@@ -1,4 +1,3 @@
 .DS_Store
 data
-message-processor/__pycache__/
-rss-feed-parser/__pycache__/
+__pycache__/
